+++ examples/counter/src/main.rs
@@ -20,7 +20,7 @@ impl Counter
 
 //-------------------------------------------------------------------------------------------------------------------
 
-fn build_ui(mut c: Commands, mut s: ResMut<SceneBuilder>)
+fn build_ui(mut c: Commands, mut s: SceneBuilder)
 {
     let scene = ("main.cob", "root");
     c.ui_root().spawn_scene_and_edit(scene, &mut s, |h| {