+++ examples/cursors/src/main.rs
@@ -6,7 +6,7 @@ use bevy_cobweb_ui::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
 
-fn build_ui(mut c: Commands, mut s: ResMut<SceneBuilder>)
+fn build_ui(mut c: Commands, mut s: SceneBuilder)
 {
     c.ui_root().spawn_scene(("main.cob", "scene"), &mut s);
 }