+++ examples/localization/src/main.rs
@@ -7,7 +7,7 @@ use bevy_cobweb_ui::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
 
-fn build_ui(mut c: Commands, mut s: ResMut<SceneBuilder>)
+fn build_ui(mut c: Commands, mut s: SceneBuilder)
 {
     let scene = ("localization", "root");
     c.ui_root().spawn_scene_and_edit(scene, &mut s, |h| {
@@ -24,7 +24,7 @@ fn build_ui(mut c: Commands, mut s: ResMut<SceneBuilder>)
                     move |//
                         id: UpdateId,
                         mut c: Commands,
-                        mut s: ResMut<SceneBuilder>,
+                        mut s: SceneBuilder,
                         manifest: Res<LocalizationManifest>//
                     | {
                         // Despawn existing buttons.