+++ examples/tooltip/src/main.rs
@@ -5,7 +5,7 @@ use bevy_cobweb_ui::prelude::*;
 
 //-------------------------------------------------------------------------------------------------------------------
 
-fn build_ui(mut c: Commands, mut s: ResMut<SceneBuilder>)
+fn build_ui(mut c: Commands, mut s: SceneBuilder)
 {
     c.spawn(Camera2d);
     c.ui_root().spawn_scene(("main.cob", "scene"), &mut s);