+++ src/loading/cache/plugin.rs
@@ -69,7 +69,7 @@ fn process_cobweb_asset_files(
     mut c: Commands,
     mut commands_buffer: ResMut<CommandsBuffer>,
     mut scene_buffer: ResMut<SceneBuffer>,
-    mut scene_loader: ResMut<SceneBuilder>,
+    mut scene_builder: SceneBuilder,
 )
 {
     let type_registry = types.read();
@@ -80,7 +80,7 @@ fn process_cobweb_asset_files(
         &mut c,
         &mut commands_buffer,
         &mut scene_buffer,
-        &mut scene_loader,
+        &mut scene_builder,
     ) {
         c.react().broadcast(CobCacheUpdated);
     }
@@ -126,7 +126,7 @@ fn apply_pending_node_updates_extract(
     mut c: Commands,
     commands_buffer: Res<CommandsBuffer>,
     mut scene_buffer: ResMut<SceneBuffer>,
-    mut scene_loader: ResMut<SceneBuilder>,
+    mut scene_builder: SceneBuilder,
     #[cfg(feature = "editor")] mut editor: ResMut<crate::editor::CobEditor>,
 )
 {
@@ -142,7 +142,7 @@ fn apply_pending_node_updates_extract(
         &loadables,
         &mut c,
         &mut scene_buffer,
-        &mut scene_loader,
+        &mut scene_builder,
         #[cfg(feature = "editor")]
         &mut editor,
     );
@@ -178,12 +178,12 @@ fn cleanup_despawned_loaded_entities(world: &mut World)
     // the RemovedComponents state will stay the same.
     fn cleanup_fn(
         mut scene_buffer: ResMut<SceneBuffer>,
-        mut scene_loader: ResMut<SceneBuilder>,
+        mut scene_builder: SceneBuilder,
         mut removed: RemovedComponents<HasLoadables>,
     )
     {
         for removed in removed.read() {
-            scene_buffer.remove_entity(&mut scene_loader, removed);
+            scene_buffer.remove_entity(&mut scene_builder, removed);
         }
     }